+++ tests/integration/utils/test_consumers.py
@@ -1,3 +1,4 @@
+import asyncio
 from importlib import import_module
 
 import pytest
@@ -43,15 +44,37 @@ async def prepare_element_cache(settings):
 
 
 @pytest.fixture
-async def communicator(request, event_loop):
-    communicator = WebsocketCommunicator(application, "/ws/site/")
-    yield communicator
-    await communicator.disconnect()
+async def get_communicator():
+    communicator: WebsocketCommunicator = None
+
+    def get_communicator(query_string=''):
+        nonlocal communicator  # use the outer communicator variable
+        if query_string:
+            query_string = "?{}".format(query_string)
+        communicator = WebsocketCommunicator(application, "/ws/site/{}".format(query_string))
+        return communicator
+
+    yield get_communicator
+    if communicator:
+        await communicator.disconnect()
+
+
+@pytest.fixture
+async def communicator(get_communicator):
+    yield get_communicator()
 
 
 @pytest.mark.asyncio
-async def test_normal_connection(communicator):
+async def test_normal_connection(get_communicator):
     await set_config('general_system_enable_anonymous', True)
+    connected, __ = await get_communicator().connect()
+    assert connected
+
+
+@pytest.mark.asyncio
+async def test_connection_with_change_id(get_communicator):
+    await set_config('general_system_enable_anonymous', True)
+    communicator = get_communicator('change_id=0')
     await communicator.connect()
 
     response = await communicator.receive_json_from()
@@ -61,7 +84,8 @@ async def test_normal_connection(communicator):
     assert type == 'autoupdate'
     assert 'changed' in content
     assert 'deleted' in content
-    assert 'change_id' in content
+    assert 'from_change_id' in content
+    assert 'to_change_id' in content
     assert Collection1().get_collection_string() in content['changed']
     assert Collection2().get_collection_string() in content['changed']
     assert TConfig().get_collection_string() in content['changed']
@@ -69,12 +93,31 @@ async def test_normal_connection(communicator):
 
 
 @pytest.mark.asyncio
-async def test_receive_changed_data(communicator):
+async def test_connection_with_invalid_change_id(get_communicator):
+    await set_config('general_system_enable_anonymous', True)
+    communicator = get_communicator('change_id=invalid')
+    connected, __ = await communicator.connect()
+
+    assert connected is False
+
+
+@pytest.mark.asyncio
+async def test_changed_data_autoupdate_off(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    await communicator.receive_json_from()
 
-    # Change a config value after the startup data has been received
+    # Change a config value
+    await set_config('general_event_name', 'Test Event')
+    assert await communicator.receive_nothing()
+
+
+@pytest.mark.asyncio
+async def test_changed_data_autoupdate_on(get_communicator):
+    await set_config('general_system_enable_anonymous', True)
+    communicator = get_communicator('autoupdate=on')
+    await communicator.connect()
+
+    # Change a config value
     await set_config('general_event_name', 'Test Event')
     response = await communicator.receive_json_from()
 
@@ -114,10 +157,10 @@ async def test_with_user():
 
 
 @pytest.mark.asyncio
-async def test_receive_deleted_data(communicator):
+async def test_receive_deleted_data(get_communicator):
     await set_config('general_system_enable_anonymous', True)
+    communicator = get_communicator('autoupdate=on')
     await communicator.connect()
-    await communicator.receive_json_from()
 
     # Delete test element
     await sync_to_async(inform_deleted_data)([(Collection1().get_collection_string(), 1)])
@@ -133,8 +176,6 @@ async def test_receive_deleted_data(communicator):
 async def test_send_invalid_notify_not_a_list(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to({'type': 'notify', 'content': {'testmessage': 'foobar, what else.'}, 'id': 'test_send_invalid_notify_not_a_list'})
     response = await communicator.receive_json_from()
@@ -148,8 +189,6 @@ async def test_send_invalid_notify_not_a_list(communicator):
 async def test_send_invalid_notify_no_elements(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to({'type': 'notify', 'content': [], 'id': 'test_send_invalid_notify_no_elements'})
     response = await communicator.receive_json_from()
@@ -163,8 +202,6 @@ async def test_send_invalid_notify_no_elements(communicator):
 async def test_send_invalid_notify_str_in_list(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to({'type': 'notify', 'content': [{}, 'testmessage'], 'id': 'test_send_invalid_notify_str_in_list'})
     response = await communicator.receive_json_from()
@@ -178,8 +215,6 @@ async def test_send_invalid_notify_str_in_list(communicator):
 async def test_send_valid_notify(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to({'type': 'notify', 'content': [{'testmessage': 'foobar, what else.'}], 'id': 'test'})
     response = await communicator.receive_json_from()
@@ -196,8 +231,6 @@ async def test_send_valid_notify(communicator):
 async def test_invalid_websocket_message_type(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to([])
 
@@ -209,8 +242,6 @@ async def test_invalid_websocket_message_type(communicator):
 async def test_invalid_websocket_message_no_id(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to({'type': 'test', 'content': 'foobar'})
 
@@ -222,8 +253,6 @@ async def test_invalid_websocket_message_no_id(communicator):
 async def test_invalid_websocket_message_no_content(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to({'type': 'test', 'id': 'test_id'})
 
@@ -235,8 +264,6 @@ async def test_invalid_websocket_message_no_content(communicator):
 async def test_send_unknown_type(communicator):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to({'type': 'if_you_add_this_type_to_openslides_I_will_be_sad', 'content': True, 'id': 'test_id'})
 
@@ -249,8 +276,6 @@ async def test_send_unknown_type(communicator):
 async def test_request_constants(communicator, settings):
     await set_config('general_system_enable_anonymous', True)
     await communicator.connect()
-    # Await the startup data
-    await communicator.receive_json_from()
 
     await communicator.send_json_to({'type': 'constants', 'content': '', 'id': 'test_id'})
 
@@ -258,3 +283,96 @@ async def test_request_constants(communicator, settings):
     assert response['type'] == 'constants'
     # See conftest.py for the content of 'content'
     assert response['content'] == {'constant1': 'value1', 'constant2': 'value2'}
+
+
+@pytest.mark.asyncio
+async def test_send_get_elements(communicator):
+    await set_config('general_system_enable_anonymous', True)
+    await communicator.connect()
+
+    await communicator.send_json_to({'type': 'getElements', 'content': {}, 'id': 'test_id'})
+    response = await communicator.receive_json_from()
+
+    type = response.get('type')
+    content = response.get('content')
+    assert type == 'autoupdate'
+    assert 'changed' in content
+    assert 'deleted' in content
+    assert 'from_change_id' in content
+    assert 'to_change_id' in content
+    assert Collection1().get_collection_string() in content['changed']
+    assert Collection2().get_collection_string() in content['changed']
+    assert TConfig().get_collection_string() in content['changed']
+    assert TUser().get_collection_string() in content['changed']
+
+
+@pytest.mark.asyncio
+async def test_send_get_elements_to_big_change_id(communicator):
+    await set_config('general_system_enable_anonymous', True)
+    await communicator.connect()
+
+    await communicator.send_json_to({'type': 'getElements', 'content': {'change_id': 1_000_000_000_000}, 'id': 'test_id'})
+    response = await communicator.receive_json_from()
+
+    type = response.get('type')
+    assert type == 'error'
+    assert response.get('in_response') == 'test_id'
+
+
+@pytest.mark.asyncio
+async def test_send_get_elements_to_small_change_id(communicator):
+    await set_config('general_system_enable_anonymous', True)
+    await communicator.connect()
+
+    await communicator.send_json_to({'type': 'getElements', 'content': {'change_id': 1}, 'id': 'test_id'})
+    response = await communicator.receive_json_from()
+
+    type = response.get('type')
+    assert type == 'autoupdate'
+    assert response.get('in_response') == 'test_id'
+    assert response.get('content')['all_data']
+
+
+@pytest.mark.asyncio
+async def test_send_invalid_get_elements(communicator):
+    await set_config('general_system_enable_anonymous', True)
+    await communicator.connect()
+
+    await communicator.send_json_to({'type': 'getElements', 'content': {'change_id': 'some value'}, 'id': 'test_id'})
+    response = await communicator.receive_json_from()
+
+    type = response.get('type')
+    assert type == 'error'
+    assert response.get('in_response') == 'test_id'
+
+
+@pytest.mark.asyncio
+async def test_turn_on_autoupdate(communicator):
+    await set_config('general_system_enable_anonymous', True)
+    await communicator.connect()
+
+    await communicator.send_json_to({'type': 'autoupdate', 'content': 'on', 'id': 'test_id'})
+    await asyncio.sleep(0.01)
+    # Change a config value
+    await set_config('general_event_name', 'Test Event')
+    response = await communicator.receive_json_from()
+
+    id = config.get_key_to_id()['general_event_name']
+    type = response.get('type')
+    content = response.get('content')
+    assert type == 'autoupdate'
+    assert content['changed'] == {
+        'core/config': [{'id': id, 'key': 'general_event_name', 'value': 'Test Event'}]}
+
+
+@pytest.mark.asyncio
+async def test_turn_off_autoupdate(get_communicator):
+    await set_config('general_system_enable_anonymous', True)
+    communicator = get_communicator('autoupdate=on')
+    await communicator.connect()
+
+    await communicator.send_json_to({'type': 'autoupdate', 'content': False, 'id': 'test_id'})
+    await asyncio.sleep(0.01)
+    # Change a config value
+    await set_config('general_event_name', 'Test Event')
+    assert await communicator.receive_nothing()