+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Core:
  - Enabled docs for using OpenSlides with Gunicorn and Uvicorn in big
    mode [#3799, #3817].
  - Changed format for elements send via autoupdate [#3926].
+ - Add a change-id system to get only new elements [#3938].
 
 Motions:
  - Option to customly sort motions [#3894].